chore: refactoring Timeline
